requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
newsapi-python==0.2.7
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.32.0
plotly==5.19.0
tqdm>=4.66.0 